backend/requirements.txt
# ============================================================
# PageTutor AI - Backend Requirements
# Author: Mustakim Shaikh (https://github.com/MustakimShaikh01)
# ============================================================

# --- Web Framework ---
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# --- Authentication & Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.1
httpx==0.27.0
itsdangerous==2.2.0

# --- Database (PostgreSQL) ---
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# --- Redis (Caching + Rate Limiting + Queue) ---
redis==5.0.4
celery[redis]==5.4.0
flower==2.0.1

# --- Object Storage (S3 / MinIO) ---
boto3==1.34.120
aioboto3==13.1.1

# --- Vector Database ---
qdrant-client==1.9.1

# --- PDF Processing ---
pypdf2==3.0.1
pdfplumber==0.11.0
pymupdf==1.24.3

# --- AI / LLM ---
openai==1.30.5
transformers==4.41.2
torch==2.3.0
sentence-transformers==3.0.1
tiktoken==0.7.0

# --- TTS (Text-to-Speech) ---
TTS==0.22.0
pyttsx3==2.90

# --- Video / Media Processing ---
moviepy==1.0.3
Pillow==10.3.0
python-pptx==0.6.23

# --- API Documentation ---
pydantic==2.7.2
pydantic-settings==2.3.0

# --- HTTP Client ---
aiohttp==3.9.5
requests==2.32.3

# --- File Hashing & Validation ---
python-magic==0.4.27
hashlib-compat==1.0.0

# --- Monitoring ---
prometheus-client==0.20.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0

# --- Logging ---
structlog==24.2.0
python-json-logger==2.0.7

# --- Email ---
fastapi-mail==1.4.1

# --- Utilities ---
python-dotenv==1.0.1
PyYAML==6.0.1
orjson==3.10.3
ujson==5.10.0
croniter==2.0.5
humanize==4.9.0
slugify==0.0.1
python-slugify==8.0.4
